requirements.txt
Flask
flasgger
PyYAML
dicttoxml
gunicorn
